Created a workaround for CUICatalog not finding 3x assets
For some reason, when -[CUICatalog
imageWithName:scaleFactor:deviceIdiom:] is called with scaleFactor of
3.0 and iPhone idiom, it returns image with scale factor of 2.0 even
through asset catalog contains image with scaleFactor of 3.0.
 Looping through all images in the catalog manually allows us to find
the correct image with correct scale.

This loop doesn't check for device idiom and it has been only tested
with assets for iPhone. It should be extended to work with other device
idioms in the future.